move version to .version file for easier import with app versions
