+++ pyzabbix/__init__.py
@@ -1,4 +1,2 @@
 from .api import ZabbixAPI, ZabbixAPIException, ssl_context_compat
 from .sender import ZabbixMetric, ZabbixSender, ZabbixResponse
-
-__version__ = '1.1.5'
